requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anthropic
anyio
async-timeout
attrs
banal
boto3
botocore
certifi
charset-normalizer
datasets
dill
distro
exceptiongroup
filelock
frozenlist
fsspec
greenlet
h11
httpcore
httpx
huggingface-hub
idna
jiter
jmespath
Mako
MarkupSafe
MouseInfo
multidict
multiprocess
numpy
openai
packaging
pandas
pillow
playwright
propcache
psutil
pyarrow
PyAutoGUI
pydantic
pydantic_core
pyee
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dateutil
python3-xlib
pytweening
pytz
PyYAML
requests
s3transfer
six
sniffio
SQLAlchemy
tenacity
tqdm
typing_extensions
tzdata
urllib3
xxhash
yarl
